requirements.txt
PyOpenGL>=3.1.5
glfw>=1.12.0
opencv-python>=4.5.1.48
numpy>=1.19.5
pywin32>=300